refact: add attachCopySource and attachMarkList call functions
