+++ demo/api.ts
@@ -23,3 +23,5 @@ export const addMark = async (params) => {
   const jsonRes = await res.json();
   return jsonRes;
 };
+
+// 获取
