+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@copus/editor",
-  "version": "0.16.0-8",
+  "version": "0.16.0-9.b2",
   "type": "module",
   "publishConfig": {
     "access": "public"
